% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{convCheck.relative}
\alias{convCheck.relative}
\title{Relative difference convergence check criterion}
\usage{
convCheck.relative(Wnew, Wold)
}
\arguments{
\item{Wnew}{Current weight matrix}

\item{Wold}{Weight matrix from the previous iteration}
}
\value{
Scalar value of the convergence criterion
}
\description{
Compares two weight matrices and returns the value of the relative 
difference convergence criterion.
}
\details{
Relative difference convergence criterion is the maximum of relative differences between
weights from two iterations
}
\seealso{
Other Convergence checks: \code{\link{convCheck.absolute}},
  \code{\link{convCheck.square}}
}

