% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{convCheck.square}
\alias{convCheck.square}
\title{Squared difference convergence check criterion}
\usage{
convCheck.square(Wnew, Wold)
}
\arguments{
\item{Wnew}{Current weight matrix}

\item{Wold}{Weight matrix from the previous iteration}
}
\value{
Scalar value of the convergence criterion
}
\description{
Compares two weight matrices and returns the value of the square difference
convergence criterion.
}
\details{
Squared difference convergence criterion is the maximum of squared absolute
differences between weights from two iterations
}
\seealso{
Other Convergence checks: \code{\link{convCheck.absolute}};
  \code{\link{convCheck.relative}}
}

