% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.convergenceCheck.R
\name{convergenceCheck}
\alias{convCheck.absolute}
\alias{convCheck.relative}
\alias{convCheck.square}
\alias{convergenceCheck}
\title{Convergence checks}
\usage{
convCheck.relative(Wnew, Wold)

convCheck.square(Wnew, Wold)

convCheck.absolute(Wnew, Wold)
}
\arguments{
\item{Wnew}{Current weight matrix}

\item{Wold}{Weight matrix from the previous iteration}
}
\value{
Scalar value of the convergence criterion
}
\description{
The convergence check functions compare two weight matrices and returns the value of the
convergence criterion describing the difference between the two weight matrices.
}
\section{Functions}{
\itemize{
\item \code{convCheck.relative}: maximum of relative differences between
weights from two iterations

\item \code{convCheck.square}: maximum of squared absolute 
differences between weights from two iterations.

\item \code{convCheck.absolute}: maximum of absolute differences
between weights from two iterations
}}

