% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.postestimation.R
\name{cr}
\alias{cr}
\title{Composite Reliability indices for matrixpls results}
\usage{
cr(object, ...)
}
\arguments{
\item{object}{matrixpls estimation result object produced by the \code{\link{matrixpls}} function.}

\item{...}{All other arguments are ignored.}
}
\value{
A named numeric vector containing the Composite Reliability indices.
}
\description{
The \code{matrixpls} method for the generic function \code{cr} computes Composite Reliability 
indices for the model using the formula presented by Fornell and Larcker (1981).
}
\references{
Fornell, C., & Larcker, D. F. (1981). Evaluating structural equation models with unobservable variables and measurement error. \emph{Journal of marketing research}, 18(1), 39–50.
}
\seealso{
Other post-estimation functions: \code{\link{R2}},
  \code{\link{ave}}, \code{\link{effects.matrixpls}},
  \code{\link{fitted.matrixpls}}, \code{\link{gof}},
  \code{\link{htmt}}, \code{\link{loadings.matrixpls}},
  \code{\link{predict.matrixpls}},
  \code{\link{residuals.matrixpls}}
}

