% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.plsc.R
\name{estimator.CFALoadings}
\alias{estimator.CFALoadings}
\title{Parameter estimation with confirmatory factor analysis of the full model}
\usage{
estimator.CFALoadings(S, model, W, ...)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{There are two options for this argument: 1. lavaan script or lavaan parameter
table, or 2. a list containing three matrices
\code{inner}, \code{reflective}, and \code{formative} defining the free regression paths
in the model.}

\item{...}{All other arguments are passed through to \code{weightFunction} and \code{parameterEstimator}.}
}
\value{
A named vector of parameter estimates.
}
\description{
Parameter estimation with confirmatory factor analysis of the full model
}
\details{
Estimates a maximum likelihood confirmatory factor analysis with \code{\link[lavaan]{lavaan}}
}
\seealso{
Other parameter estimators: \code{\link{estimator.EFALoadings}};
  \code{\link{estimator.PLScLoadings}};
  \code{\link{params.plsregression}};
  \code{\link{params.regression}}
}

