% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.plsc.R
\name{estimator.EFALoadings}
\alias{estimator.EFALoadings}
\title{Parameter estimation with per-block exploratory factor analysis}
\usage{
estimator.EFALoadings(S, model, W, ..., fm = "minres")
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{There are two options for this argument: 1. lavaan script or lavaan parameter
table, or 2. a list containing three matrices
\code{inner}, \code{reflective}, and \code{formative} defining the free regression paths
in the model.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{...}{All other arguments are passed through to \code{weightFunction} and \code{parameterEstimator}.}

\item{fm}{factoring method for estimating the factor loadings. Passed through to \code{\link[psych]{fa}}.}
}
\value{
A named vector of parameter estimates.
}
\description{
Estimates the factor loading parameters one latent variable at at time with exploratory
factor analysis using the \code{\link[psych]{fa}} function from the \code{psych} package.
}
\details{
Estimating an unconstrained single factor model requires three indicators. The loadings of 
single indicator factors are estimated as 1 and two indicator factors as estimated by the
square root of the indicator correlation.
}
\seealso{
Other parameter estimators: \code{\link{estimator.CFALoadings}},
  \code{\link{estimator.PLScLoadings}},
  \code{\link{params.regression}}
}

