% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{estimator.ols}
\alias{estimator.ols}
\title{Parameter estimation with separate regression analyses}
\usage{
estimator.ols(S, model, W, ..., C = NULL, IC = NULL)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{A model matrix with dependent variables on rows, independent variables on colums, and
non-zero elements indicating relationships. Can be either \code{inner}, \code{reflective},
or \code{formative} matrix.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{...}{All other arguments are ignored.}

\item{C}{Correlation matrix of the composites.}

\item{IC}{Correlation matrix of the composites and indicators.}
}
\value{
A matrix with estimated parameters.
}
\description{
Estimates the parameters of a model matrix with OLS regression.
}
\details{
Providing \code{C} or \code{IC} allows for using disattenuated or otherwise
adjusted correlation matrices. If not provided, these matrices are calculated using \code{S} and
\code{W}.
}
\seealso{
Other regression estimators: \code{\link{estimator.plsreg}},
  \code{\link{estimator.tsls}}
}

