% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.plspm.R
\name{estimator.plsreg}
\alias{estimator.plsreg}
\title{Parameter estimation with PLS regressions}
\usage{
estimator.plsreg(S, model, W, ..., C)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{A model matrix with dependent variables on rows, independent variables on colums, and
non-zero elements indicating relationships. Onyl the \code{inner} matrix can be used.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{...}{All other arguments are ignored.}

\item{C}{Correlation matrix of the composites.}
}
\value{
A matrix with estimated parameters.
}
\description{
Estimates the parameters of a model matrix (\code{inner}) with PLS least squares regressions.
}
\references{
Sanchez, G. (2013). \emph{PLS Path Modeling with R.} Retrieved from http://www.gastonsanchez.com/PLS Path Modeling with R.pdf

Bjørn-Helge Mevik, & Ron Wehrens. (2007). The pls Package:  Principal Component and Partial Least Squares Regression in R. \emph{Journal of Statistical Software}, 18. Retrieved from http://www.jstatsoft.org/v18/i02/paper
}
\seealso{
Other regression estimators: \code{\link{estimator.ols}},
  \code{\link{estimator.tsls}}
}

