% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inner.GSCA}
\alias{inner.GSCA}
\title{GSCA inner estimation}
\usage{
inner.GSCA(S, W, inner.mod, ...)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{inner.mod}{A square matrix specifying the relationships of the composites in the model.}

\item{...}{Other parameters are ignored}
}
\value{
A matrix of unscaled inner weights \code{E} with the same dimesions as \code{inner.mod}.
}
\description{
This implements the first step of the GSCA estimation describe by Hwang & Takane (2004). GSCA inner estimation should
be used only with GSCA outer estimation.
}
\details{
The first step of GSCA estimation method, as describe by Hwang & Takane (2004), involves estimating all model regressions,
including also the relationships from composites toward indicators in the first step.

The implementation of GSCA in matrixpls differs from the Hwang & Takane (2004) version in that during the
first step, only regressions between composites are estimated. The reason for this is that the
relationhips from the composites to indicators are simple regressions that are simply the covariances between
the indicators and compositess. Since these covariances need to be calculated in the second step, it is more
efficient to not calculate them during the first step.

This algorithm is therefore identical to the PLS path weighting scheme with the exception that correlations
are not used for inverse relationships and there is no falling back to identity scheme for composites
that are not connected to other composites.
}
\examples{
# Run the example from plspm package using GSCA estimation

if(require(ASGSCA)) {
  
  # Run the GSCA example from the ASGSCA package
  
  #Scenario (g) in Romdhani et al. (2014): 4 SNPs mapped to 2 genes and 4 
  #traits involved in 2 clinical pathways 
  #In total: 8 observed variables and 4 latent variables.
  #One of the traits is involved in both clinical pathways.
  #One gene is connected to one of the clinical pathways and
  #the other to both of them.
  
  data(GenPhen)
  W0 <- matrix(c(rep(1,2),rep(0,8),rep(1,2),rep(0,8),rep(1,3),rep(0,7),rep(1,2)),nrow=8,ncol=4)
  B0 <- matrix(c(rep(0,8),rep(1,2),rep(0,3),1,rep(0,2)),nrow=4,ncol=4)
  
  #Estimation only
  GSCA.res <- GSCA(GenPhen,W0, B0,estim=TRUE,path.test=FALSE, latent.names=c("Gene1","Gene2","Clinical pathway 1","Clinical pathway 2"))
  
  # Setup matrixpls to estimate the same model. Note GSCA places dependent variables
  # on columns but matrixpls uses rows for dependent variables
  
  inner <- t(B0)
  reflective <- matrix(0,8,4)
  formative <- t(W0)
  
  colnames(formative) <- rownames(reflective) <- names(GenPhen)
  colnames(inner) <- rownames(inner) <- rownames(formative) <- colnames(reflective) <- c("Gene1","Gene2","Clinical pathway 1","Clinical pathway 2")
  
  model <- list(inner = inner, 
                reflective = reflective,
                formative = formative)
  
  # Estimate using direct minimization of the estimation criterion
  
  matrixpls.res1 <- matrixpls(cov(GenPhen),  model,
                              weightFunction = weight.optim,
                              optimCriterion = optim.minimizeGSCA)
  
  # Compare thw weights
  print(GSCA.res$Weight)
  print(t(attr(matrixpls.res1,"W")))
  
  # Compare the paths
  print(GSCA.res$Path)
  print(t(attr(matrixpls.res1,"beta")))
  
  
  # Estimate using alternative least squares
  
  matrixpls.res2 <- matrixpls(cov(GenPhen),  model,
                              outerEstimators = outer.GSCA,
                              innerEstimator = inner.GSCA)

  # Compare thw weights
  print(GSCA.res$Weight)
  print(t(attr(matrixpls.res2,"W")))
  
  # Compare the paths
  print(GSCA.res$Path)
  print(t(attr(matrixpls.res2,"beta")))
  
  
} else{
  print("This example requires the ASGSCA package")
}

}
\references{
Hwang, H., & Takane, Y. (2004). Generalized structured component analysis. Psychometrika, 69(1), 81–99. doi:10.1007/BF02295841
}
\seealso{
Other inner estimators: \code{\link{inner.centroid}};
  \code{\link{inner.factor}}; \code{\link{inner.identity}};
  \code{\link{inner.path}}
}

