% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadings}
\alias{loadings}
\title{Factor loadings matrix from matrixpls results}
\usage{
loadings(object, ...)
}
\arguments{
\item{...}{All other arguments are ignored.}

\item{x}{An object of class \code{matrixpls}}
}
\value{
A matrix of factor loadings.
}
\description{
The \code{matrixpls} method for the generic function \code{loadings} computes standardized factor loading matrix
from the model results.
}
\seealso{
Other post-estimation functions: \code{\link{AVE}};
  \code{\link{CR}}; \code{\link{GoF}}; \code{\link{R2}};
  \code{\link{effects.matrixpls}};
  \code{\link{predict.matrixpls}};
  \code{\link{residuals.matrixpls}}
}

