% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrixpls.internal}
\alias{matrixpls.internal}
\title{Partial Least Squares estimation - Internal}
\usage{
matrixpls.internal(S, model, W.mod = NULL,
  parameterEstimator = params.regression, outerEstimators = NULL,
  innerEstimator = inner.path, weightEstimator = NULL, method = NULL, ...,
  convCheck = function(W, W_old) {     max(abs(W_old - W)) }, tol = 1e-05,
  iter = 100, validateInput = TRUE, standardize = TRUE)
}
\arguments{
\item{...}{All other arguments are passed through to \code{matrixpls.weights} or
\code{\link{matrixpls.optim}} and \code{parameterEstimator}.}

\item{S}{Covariance matrix of the data.}

\item{model}{There are two options for this argument: 1. lavaan script or lavaan parameter
table, or 2. a list containing three matrices
\code{inner}, \code{reflective}, and \code{formative} defining the free regression paths
in the model.}

\item{W.mod}{An optional numeric matrix representing the weight model and starting weights
(i.e. the how the indicators are combined to form the composite variables). If this argument is not specified,
the weight model is defined based on the relationships in the \code{reflective} and  \code{formative}
elements of \code{model}.}

\item{parameterEstimator}{A function that takes three arguments, the data covariance matrix \code{S},
weights \code{W}, and model specification \code{model} and returns a named vector of parameter estimates. The default is \code{\link{params.regression}}}

\item{outerEstimators}{A function or a list of functions used for outer estimation. If
the value of this parameter is a function, the same function is applied to all
composites. If the value of the parameter is a list, the composite \code{n} is estimated
with the estimator in the \code{n}th position in the list. If this argument is
\code{NULL} the starting weights specified in \code{W.mod} will be returned. The default is \code{\link{outer.modeA}}
(PLS Mode A estimation).}

\item{innerEstimator}{A function used for inner estimation. Setting this argument to
\code{null} will use identity matrix as inner estimates and causes the algorithm to converge
after the first iteration. This is useful when using
\code{\link{outer.fixedWeights}} or some other outer estimation function that
does not use inner estimation results. The default is \code{\link{inner.path}}
(PLS path weighting scheme).}

\item{convCheck}{A function that takes the old and new weight matrices and
returns a scalar that is compared against \code{tol} to check for convergence. The default
function calculates the differences between each cell of old and new weights and returns the largest
absolute difference.}

\item{tol}{Decimal value indicating the tolerance criterion for the
iterations.}

\item{iter}{An integer indicating the maximum number of iterations.}

\item{validateInput}{A boolean indicating whether the validity of the parameter values should be tested.}

\item{standardize}{\code{TRUE} (default) or \code{FALSE} indicating whether S should be converted
to correlation matrix.}

\item{weightEstimator}{A function that taking an object of class class
\code{matrixpls} and returning a scalar. The default is \code{\link{optim.maximizeInnerR2}}}

\item{method}{The minimization algorithm to be used. See \code{\link[stats]{optim}}
for details. Default is "BFGS".}
}
\value{
A named numeric vector of class \code{matrixpls} containing parameter estimates followed by weights.
}
\description{
Implementation of \code{\link{matrixpls}} and \code{\link{matrixpls.optim}}
}
\details{
See \code{\link{matrixpls}} and \code{\link{matrixpls.optim}} for details.
}
\seealso{
\code{\link{matrixpls}} and \code{\link{matrixpls.optim}}
}

