% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{optim.gsca}
\alias{optim.gsca}
\title{GSCA optimization criterion}
\usage{
optim.gsca(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Sum of residual variances.
}
\description{
Optimization criterion for minimizing the sum of all residual
variances in the model. 

The matrixpls implementation of the GSCA criterion extends the criterion
presented by Huang and Takane by including also the minimization of the
residual variances of the formative part of the model. The formative
regressions in a model are typically specified to be identical to the 
weight pattern \code{W.model} resulting zero residual variances by definition.
However, it is possible to specify a formative model that does not completely
overlap with the weight pattern leading to non-zero residuals that can be
optimizes.
}
\references{
Hwang, H., & Takane, Y. (2004). Generalized structured component analysis. 
Psychometrika, 69(1), 81–99. doi:10.1007/BF02295841
}
\seealso{
Other GSCA functions: \code{\link{inner.gsca}},
  \code{\link{outer.gsca}}

Other Weight optimization criteria: \code{\link{optim.maximizeFullR2}},
  \code{\link{optim.maximizeIndicatorR2}},
  \code{\link{optim.maximizeInnerR2}}
}

