% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{optim.maximizeFullR2}
\alias{optim.maximizeFullR2}
\title{Optimization criterion for maximizing reflective and innner R2}
\usage{
optim.maximizeFullR2(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Negative of sum of R2 values of the \code{inner} and
\code{reflective} matrices.
}
\description{
Calculates the sum of R2s of \code{inner} matrix and 
\code{reflective} matrix. The
optimization criterion is negative of this sum.
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.gsca}},
  \code{\link{optim.maximizeIndicatorR2}},
  \code{\link{optim.maximizeInnerR2}}
}

