% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{optim.maximizeIndicatorR2}
\alias{optim.maximizeIndicatorR2}
\title{Optimization criterion for maximal prediction}
\usage{
optim.maximizeIndicatorR2(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Mean squared prediction error.
}
\description{
Optimization criterion for maximal prediction
}
\details{
Calculates the predicted variances of reflective indicators. The
prediction criterion is negative of the sum of predicted variances.
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.GSCA}};
  \code{\link{optim.maximizeFullR2}};
  \code{\link{optim.maximizeInnerR2}};
  \code{\link{optim.maximizePrediction}}
}

