% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optim.minimizeGSCA}
\alias{optim.minimizeGSCA}
\title{GSCA optimization criterion}
\usage{
optim.minimizeGSCA(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Sum of residual variances.
}
\description{
GSCA optimization criterion
}
\details{
Optimization criterion for minimizing the sum of all residual
variances in the model
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.maximizeInnerR2}};
  \code{\link{optim.maximizePrediction}}
}

