% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outer.fixedWeights}
\alias{outer.fixedWeights}
\title{PLS outer estimation with fixed weights}
\usage{
outer.fixedWeights(S, W, E, W.mod, ...)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{E}{Inner weight matrix. A square matrix of inner estimates between the composites.}

\item{W.mod}{A matrix specifying the weight relationships and their starting values.}

\item{...}{Other parameters are ignored}
}
\value{
A matrix of unscaled outer weights \code{W} with the same dimesions as \code{W.mod}.
}
\description{
Returns the starting weights as outer estimates.
}
\seealso{
Other outer estimators: \code{\link{outer.GSCA}};
  \code{\link{outer.RGCCA}}; \code{\link{outer.factor}};
  \code{\link{outer.modeA}}; \code{\link{outer.modeB}}
}

