% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outer.GSCA}
\alias{outer.GSCA}
\title{GSCA outer estimation}
\usage{
outer.GSCA(S, W, E, W.mod, ...)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{E}{Inner weight matrix. A square matrix of inner estimates between the composites.}

\item{W.mod}{A matrix specifying the weight relationships and their starting values.}

\item{...}{Other parameters are ignored}
}
\value{
A matrix of unscaled outer weights \code{W} with the same dimesions as \code{W.mod}.
}
\description{
This implements the second step of the GSCA estimation describe by Hwang & Takane (2004). GSCA outer estimation should
be used only with GSCA inner estimation.
}
\details{
The second step of GSCA estimation method, as describe by Hwang & Takane (2004), involves calculation of new
weights given the regression estimates form the first step. In the second step, the following function is
minimized (Hwang & Takane, 2004, eq. 7, first row):

\deqn{SS(Z[V-\Lambda])}{SS(Z[V-Lambda])}

Because \eqn{\Lambda}{Lambda} is defined as \eqn{WA}{WA}, the function to be minimized is
identical to the first step function (Hwang & Takane, 2004, eq. 4, first row):

\deqn{SS(ZV-ZWA)}{SS(ZV-ZWA)}

In the second step, this function is minimized in respect to weights \eqn{W}{W} and \eqn{V}{V}.
This involves estimating each regression ananalysis in the model including
regressions between the composites and from composites to indicators and to minimize the sum of all OLS
dicrepancy functions simultaneously. Because one weight can be included in many regressions, these equations
must be estimated simultaneously. The minimization algoritm is the Nelder-Mead algorithm implemented in the
\code{\link{optim}} function.

The GSCA algoritm described by Hwang & Takane (2004) allows some indicators to be excluded from the second
step, but in this implementation all indicators are always used so that each weight relation described in
\code{W.mod} has always a corresponding regression relationship from a composite to a variable in the second
step of GSCA estimation.
}
\examples{
# Run the example from plspm package using GSCA estimation

if(require(ASGSCA)) {

  # Run the GSCA example from the ASGSCA package
  
  #Scenario (g) in Romdhani et al. (2014): 4 SNPs mapped to 2 genes and 4 
  #traits involved in 2 clinical pathways 
  #In total: 8 observed variables and 4 latent variables.
  #One of the traits is involved in both clinical pathways.
  #One gene is connected to one of the clinical pathways and
  #the other to both of them.
  
  data(GenPhen)
  W0 <- matrix(c(rep(1,2),rep(0,8),rep(1,2),rep(0,8),rep(1,3),rep(0,7),rep(1,2)),nrow=8,ncol=4)
  B0 <- matrix(c(rep(0,8),rep(1,2),rep(0,3),1,rep(0,2)),nrow=4,ncol=4)
  
  #Estimation only
  GSCA.res <- GSCA(GenPhen,W0, B0,estim=TRUE,path.test=FALSE, latent.names=c("Gene1","Gene2","Clinical pathway 1","Clinical pathway 2"))
  
  # Setup matrixpls to estimate the same model. Note GSCA places dependent variables
  # on columns but matrixpls uses rows for dependent variables
  
  inner <- t(B0)
  reflective <- W0
  reflective[,1:2] <- 0
  formative <- t(W0)
  formative[3:4,] <- 0

  colnames(formative) <- rownames(reflective) <- names(GenPhen)
  colnames(inner) <- rownames(inner) <- rownames(formative) <- colnames(reflective) <- c("Gene1","Gene2","Clinical pathway 1","Clinical pathway 2")
    
  model <- list(inner = inner, 
                reflective = reflective,
                formative = formative)

  # Estimate using direct minimization of the estimation criterion
  
  matrixpls.res1 <- matrixpls(cov(GenPhen),  model,
                              weightFunction = weight.optim,
                              optimCriterion = optim.minimizeGSCA)

  stop()
  
  # Estimate using alternative least squares

  matrixpls.res2 <- matrixpls(cov(GenPhen),  model,
            outerEstimators = outer.GSCA,
            innerEstimator = inner.GSCA)
  
} else{
  print("This example requires the ASGSCA package")
}

}
\references{
Hwang, H., & Takane, Y. (2004). Generalized structured component analysis. \emph{Psychometrika}, 69(1), 81–99. doi:10.1007/BF02295841
}
\seealso{
Other outer estimators: \code{\link{outer.factor}};
  \code{\link{outer.fixedWeights}};
  \code{\link{outer.modeA}}; \code{\link{outer.modeB}}
}

