% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.postestimation.R
\name{r2}
\alias{r2}
\title{R2	for matrixpls results}
\usage{
r2(object, ...)
}
\arguments{
\item{object}{matrixpls estimation result object produced by the \code{\link{matrixpls}} function.}

\item{...}{All other arguments are ignored.}
}
\value{
A named numeric vector containing the R2 values.
}
\description{
The \code{matrixpls} method for the generic function \code{r2} computes the squared multiple correlation (R2)
for composites predicted by other composites in the model.
}
\seealso{
Other post-estimation functions: 
\code{\link{ave}()},
\code{\link{cei}()},
\code{\link{cr}()},
\code{\link{effects.matrixpls}()},
\code{\link{fitSummary}()},
\code{\link{fitted.matrixpls}()},
\code{\link{gof}()},
\code{\link{htmt}()},
\code{\link{loadings}()},
\code{\link{predict.matrixpls}()},
\code{\link{residuals.matrixpls}()}
}
\concept{post-estimation functions}
