% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{reliability.weightLoadingProduct}
\alias{reliability.weightLoadingProduct}
\title{Reliabilities as products of weights and loadings}
\usage{
reliability.weightLoadingProduct(S, loadings, W, ...)
}
\arguments{
\item{S}{the data covariance matrix}

\item{loadings}{matrix of factor loading estimates}

\item{weights}{matrix of weights}
}
\value{
a named vector of estimated composite reliabilities.
}
\description{
Calculates reliabilities as a matrix product of loadings and weidhts
}

