% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilityEstim.R
\name{reliabilityEstim}
\alias{reliabilityEstim}
\alias{reliabilityEstim.alpha}
\alias{reliabilityEstim.weightLoadingProduct}
\title{Reliabilities as products of weights and loadings}
\usage{
reliabilityEstim.weightLoadingProduct(S, loadings, W, ...)

reliabilityEstim.alpha(S, loadings, W, ...)
}
\arguments{
\item{S}{the data covariance matrix}

\item{loadings}{matrix of factor loading estimates}

\item{W}{matrix of weights}

\item{...}{All other arguments are ignored.}
}
\value{
a named vector of estimated composite reliabilities.
}
\description{
Calculates reliabilities as a matrix product of loadings and weights.
}
\section{Functions}{
\itemize{
\item \code{reliabilityEstim.weightLoadingProduct}: Reliability estimation based on weights and loadings.

\item \code{reliabilityEstim.alpha}: Reliability estimation with Cronbach's alpha
}}

