% Generated by roxygen2 (4.0.2): do not edit by hand
\name{signChange.construct}
\alias{signChange.construct}
\title{Construct level sign change correction for boostrapping}
\usage{
signChange.construct(Worig, W)
}
\arguments{
\item{Worig}{the original weight matrix}

\item{W}{a weight matrix of a bootstrap sample}
}
\value{
A weight matrix with the same dimensions as \code{W} after
}
\description{
Changes the signs of \code{W} on all rows where the sign of the
sum of the row differs between \code{Worig} and \code{W}
}
\references{
Tenenhaus, M., Esposito Vinzi, V., Chatelin, Y.-M., & Lauro, C. (2005). PLS Path Modeling.
\emph{Computational Statistics & Data Analysis}, 48(1), 159–205. doi:10.1016/j.csda.2004.03.005
}
\seealso{
\code{\link{matrixpls.boot}}

Other signChangeCorrection: \code{\link{signChange.individual}}
}

