% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.weights.R
\name{weight.factor}
\alias{weight.factor}
\title{Blockwise factor score weights}
\usage{
weight.factor(S, model, W.model = NULL, ..., fm = "minres",
  standardize = TRUE)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{Ignored by \code{weight.factor}.}

\item{W.model}{An optional numeric matrix representing the weight patter and starting weights
(i.e. the how the indicators are combined to form the composite variables). If this argument is not specified,
the weight patter is defined based on the relationships in the \code{reflective} and  \code{formative}
elements of \code{model}.}

\item{...}{All other arguments are ignored.}

\item{fm}{factoring method for estimating the common factor model. Possible values are
\code{minres}, \code{wls}, \code{gls}, \code{pa}, and \code{ml}. The parameter is passed through to
to \code{\link[psych]{fa}}.}

\item{standardize}{A boolean indicating whether \code{S} is standardized}
}
\value{
An object of class \code{"matrixplsweights"}, which is a matrix containing the weights with the following attributes:

\item{iterations}{Number of iterations performed}

\item{converged}{A boolean indicating if the algorithm converged}

\item{history}{A data.frame containing the weights for each iteration}
}
\description{
Calculates weights by estimating a common factor analysis model with a single factor for each 
indicator block and using the resulting estimates to calculate factor score weights
}
\section{Weight specification}{


The argument \code{W.model} is a (\code{l x k}) matrix that indicates
how the indicators are combined to form the composites. The original papers about
Partial Least Squares as well as all current PLS implementations define this as
\code{t(reflective) | formative}, which means that the weight patter must match the
model specified in \code{reflective} and \code{formative}. Matrixpls does not
require that \code{W.model} needs to match \code{reflective} and \code{formative}, but
accepts any numeric matrix. If this argument is not specified, all elements of \code{W.model} that
correspond to non-zero elements in the \code{reflective} or \code{formative} formative
matrices receive the value 1.
}
\seealso{
Other Weight algorithms: \code{\link{weight.fixed}},
  \code{\link{weight.optim}}, \code{\link{weight.pls}},
  \code{\link{weight.principal}}
}

