% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{weightSign.dominantIndicator}
\alias{weightSign.dominantIndicator}
\title{Dominant indicator sign ambiguity correction}
\usage{
weightSign.dominantIndicator(W, S)
}
\arguments{
\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{S}{Covariance matrix of the data.}
}
\value{
W after sign correction.
}
\description{
Adjust the signs of W so that the first indicator of each composite has positive
weight.
}
\seealso{
\code{\link{matrixpls}};

Other Sign corrections: \code{\link{weightSign.Wold1985}}
}

