% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{weightSign.Wold1985}
\alias{weightSign.Wold1985}
\title{Wold's sign ambiguity correction}
\usage{
weightSign.Wold1985(W, S)
}
\arguments{
\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{S}{Covariance matrix of the data.}
}
\value{
W after sign correction.
}
\description{
Adjust the signs of W so that the majority of the indicators are positively
correlated with the composite as proposed by Wold (1985).
}
\references{
Wold, H. (1985). Partial Least Squares. In S. Kotz & N. L. Johnson (Eds.), Encyclopedia  of
statistical sciences (Vol. 6, pp. 581–591). New York: Wiley.
}
\seealso{
\code{\link{matrixpls}};

Other Sign corrections: \code{\link{weightSign.dominantIndicator}}
}

