% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convCheck.absolute}
\alias{convCheck.absolute}
\title{Absolute difference convergence check criterion}
\usage{
convCheck.absolute(Wnew, Wold)
}
\arguments{
\item{Wnew}{Current weight matrix}

\item{Wold}{Weight matrix from the previous iteration}
}
\value{
Scalar value of the convergence criterion
}
\description{
Compares two weight matrices and returns the value of the absolute
difference convergence criterion.
}
\details{
Absolute difference convergence criterion is the maximum of absolute differences
between weights from two iterations
}
\seealso{
Other Convergence checks: \code{\link{convCheck.relative}};
  \code{\link{convCheck.square}}
}

