% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixpls.plsc.R
\name{estimator.EFALoadings}
\alias{estimator.EFALoadings}
\title{Parameter estimation with per-block exploratory factor analysis}
\usage{
estimator.EFALoadings(S, model, W, ..., fm = "minres")
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{There are two options for this argument: 1. lavaan script or lavaan parameter
table, or 2. a list containing three matrices
\code{inner}, \code{reflective}, and \code{formative} defining the free regression paths
in the model.}

\item{...}{All other arguments are passed through to \code{weightFunction} and \code{parameterEstimator}.}

\item{fm}{factoring method for estimating the factor loadings. Passed through to \code{\link[psych]{fa}}.}
}
\value{
A named vector of parameter estimates.
}
\description{
Parameter estimation with per-block exploratory factor analysis
}
\seealso{
Other parameter estimators: \code{\link{estimator.CFALoadings}};
  \code{\link{estimator.PLScLoadings}};
  \code{\link{params.plsregression}};
  \code{\link{params.regression}}
}

