% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{estimator.tsls}
\alias{estimator.tsls}
\title{Parameter estimation with two-stage least squares regressions}
\usage{
estimator.tsls(S, model, W, ..., C)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{A model matrix with dependent variables on rows, independent variables on colums, and
non-zero elements indicating relationships. Onyl the \code{inner} matrix can be used.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{...}{All other arguments are ignored.}

\item{C}{Correlation matrix of the composites.}
}
\value{
A matrix with estimated parameters.
}
\description{
Estimates the parameters of a model matrix (\code{inner}) with two-stage least squares regressions.
The exogenous variables are used as instruments.
}
\seealso{
Other regression estimators: \code{\link{estimator.ols}},
  \code{\link{estimator.plsreg}}
}

