% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.postestimation.R
\name{fitSummary}
\alias{fitSummary}
\title{Summary of model fit of PLS model}
\usage{
fitSummary(object)
}
\arguments{
\item{object}{matrixpls estimation result object produced by the \code{\link{matrixpls}} function.}

\item{...}{All other arguments are ignored.}
}
\value{
A list containing a selection of fit indices.
}
\description{
\code{fitSummary} computes a list of statistics
that are commonly used to access the overall fit of the PLS model.
}
\seealso{
Other post-estimation functions: \code{\link{ave}},
  \code{\link{cr}}, \code{\link{effects.matrixpls}},
  \code{\link{fitted.matrixpls}}, \code{\link{gof}},
  \code{\link{htmt}}, \code{\link{loadings}},
  \code{\link{predict.matrixpls}}, \code{\link{r2}},
  \code{\link{residuals.matrixpls}}
}
