% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.postestimation.R
\name{loadings.matrixpls}
\alias{loadings.matrixpls}
\title{Factor loadings matrix from matrixpls results}
\usage{
loadings.matrixpls(object, ...)
}
\arguments{
\item{object}{An object of class \code{matrixpls}}

\item{...}{All other arguments are ignored.}
}
\value{
A matrix of factor loadings.
}
\description{
The \code{matrixpls} method for the generic function \code{loadings} computes standardized factor loading matrix
from the model results.
}
\seealso{
Other post-estimation functions: \code{\link{R2}},
  \code{\link{ave}}, \code{\link{cr}},
  \code{\link{effects.matrixpls}},
  \code{\link{fitted.matrixpls}}, \code{\link{gof}},
  \code{\link{htmt}}, \code{\link{predict.matrixpls}},
  \code{\link{residuals.matrixpls}}
}

