% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.postestimation.R
\name{loadings}
\alias{loadings}
\title{Factor loadings matrix from matrixpls results}
\usage{
loadings(object, ...)
}
\arguments{
\item{object}{matrixpls estimation result object produced by the \code{\link{matrixpls}} function.}

\item{...}{All other arguments are ignored.}
}
\value{
A matrix of estimated factor loadings.
}
\description{
The \code{matrixpls} method for the generic function \code{loadings} computes standardized factor loading matrix
from the model results.
}
\seealso{
Other post-estimation functions: 
\code{\link{ave}()},
\code{\link{cr}()},
\code{\link{effects.matrixpls}()},
\code{\link{fitSummary}()},
\code{\link{fitted.matrixpls}()},
\code{\link{gof}()},
\code{\link{htmt}()},
\code{\link{predict.matrixpls}()},
\code{\link{r2}()},
\code{\link{residuals.matrixpls}()}
}
\concept{post-estimation functions}
