% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{matrixpls-common}
\alias{matrixpls-common}
\title{Commonly used arguments}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{There are two options for this argument: 1. lavaan script or lavaan parameter
table, or 2. a list containing three matrices
\code{inner}, \code{reflective}, and \code{formative} defining the free regression paths
in the model.}

\item{W}{Weight matrix, where the indicators are on colums and composites are on the rows.}

\item{C}{Correlation matrix of the composites.}

\item{IC}{Correlation matrix of the composites and indicators.}

\item{E}{Inner weight matrix. A square matrix of inner estimates between the composites.}

\item{W.model}{A matrix specifying the weight relationships and their starting values.}
}
\description{
The following describes commonly used arguments in the \pkg{matrixpls} package.
}
