% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optim.GSCA}
\alias{optim.GSCA}
\title{GSCA optimization criterion}
\usage{
optim.GSCA(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Sum of residual variances.
}
\description{
GSCA optimization criterion
}
\details{
Optimization criterion for minimizing the sum of all residual
variances in the model
}
\references{
Hwang, H., & Takane, Y. (2004). Generalized structured component analysis.
Psychometrika, 69(1), 81–99. doi:10.1007/BF02295841
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.GCCA}};
  \code{\link{optim.maximizeInnerR2}};
  \code{\link{optim.maximizePrediction}}
}

