% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{optim.maximizeInnerR2}
\alias{optim.maximizeInnerR2}
\title{Optimization criterion for maximizing inner R2}
\usage{
optim.maximizeInnerR2(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Negative of sum of R2 values of the \code{inner} matrix.
}
\description{
Calculates the sum of R2s of \code{inner} matrix. The
optimization criterion is negative of this sum.
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.gsca}},
  \code{\link{optim.maximizeFullR2}},
  \code{\link{optim.maximizeIndicatorR2}}
}

