% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optim.maximizeInnerR2}
\alias{optim.maximizeInnerR2}
\title{Optimization criterion to maximize the inner model mean R2}
\usage{
optim.maximizeInnerR2(matrixpls.res)
}
\arguments{
\item{matrixpls.res}{An object of class \code{matrixpls} from which the
criterion function is calculated}
}
\value{
Negative mean R2.
}
\description{
Optimization criterion to maximize the inner model mean R2
}
\seealso{
Other Weight optimization criteria: \code{\link{optim.GCCA}};
  \code{\link{optim.GSCA}};
  \code{\link{optim.maximizePrediction}}
}

