% Generated by roxygen2 (4.0.2): do not edit by hand
\name{R2}
\alias{R2}
\title{R2	for matrixpls results}
\usage{
R2(object, ...)
}
\arguments{
\item{object}{matrixpls estimation result object produced by the \code{\link{matrixpls}} function.}

\item{...}{All other arguments are ignored.}
}
\value{
A named numberic vector containing the R2 values.
}
\description{
The \code{matrixpls} method for the generic function \code{R2} computes the squared multiple correlation (R2)
for composites predicted by other composites in the model.
}
\seealso{
Other post-estimation functions: \code{\link{AVE}};
  \code{\link{CR}}; \code{\link{GoF}};
  \code{\link{effects.matrixpls}};
  \code{\link{loadings.matrixpls}};
  \code{\link{predict.matrixpls}};
  \code{\link{residuals.matrixpls}}
}

