% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.reliability.R
\name{reliability.weightLoadingProduct}
\alias{reliability.weightLoadingProduct}
\title{Reliabilities as products of weights and loadings}
\usage{
reliability.weightLoadingProduct(S, loadings, W, ...)
}
\arguments{
\item{S}{the data covariance matrix}

\item{loadings}{matrix of factor loading estimates}

\item{W}{matrix of weights}

\item{...}{All other arguments are ignored.}
}
\value{
a named vector of estimated composite reliabilities.
}
\description{
Calculates reliabilities as a matrix product of loadings and weights.
}

