% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{weight.fixed}
\alias{weight.fixed}
\title{Fixed weights}
\usage{
weight.fixed(S, model, W.model = NULL, ..., standardize = TRUE)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{Ignored by \code{weight.fixed}}

\item{W.model}{An optional numeric matrix representing the weight patter and starting weights
(i.e. the how the indicators are combined to form the composite variables). If this argument is not specified,
the weight patter is defined based on the relationships in the \code{reflective} and  \code{formative}
elements of \code{model}.}

\item{...}{All other arguments are ignored.}

\item{standardize}{A boolean indicating whether \code{S} is standardized}
}
\value{
An object of class \code{"matrixplsweights"}, which is a matrix containing the weights with the following attributes:

\item{iterations}{Number of iterations performed}

\item{converged}{A boolean indicating if the algorithm converged}

\item{history}{A data.frame containing the weights for each iteration}
}
\description{
Returns fixed weights
}
\details{
Returns the starting weights specified
in \code{W.model}. If \code{standardize} is \code{TRUE} the weights are
standardized so that the composites have unit variances.
}
\seealso{
Other Weight algorithms: \code{\link{weight.factor}},
  \code{\link{weight.optim}}, \code{\link{weight.pls}},
  \code{\link{weight.principal}}
}

