% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixpls.R
\name{weight.principal}
\alias{weight.principal}
\title{Blockwise principal component weights}
\usage{
weight.principal(S, model, W.model = NULL, ..., standardize = TRUE)
}
\arguments{
\item{S}{Covariance matrix of the data.}

\item{model}{Ignored by \code{weight.principal}}

\item{W.model}{An optional numeric matrix representing the weight patter and starting weights
(i.e. the how the indicators are combined to form the composite variables). If this argument is not specified,
the weight patter is defined based on the relationships in the \code{reflective} and  \code{formative}
elements of \code{model}.}

\item{...}{All other arguments are ignored.}

\item{standardize}{A boolean indicating whether \code{S} is standardized}
}
\value{
An object of class \code{"matrixplsweights"}, which is a matrix containing the weights with the following attributes:

\item{iterations}{Number of iterations performed}

\item{converged}{A boolean indicating if the algorithm converged}

\item{history}{A data.frame containing the weights for each iteration}
}
\description{
Calculates weights by calculating a principal component analysis for each 
indicator block and returning the weights for the first principal component.
}
\seealso{
Other Weight algorithms: \code{\link{weight.factor}},
  \code{\link{weight.fixed}}, \code{\link{weight.optim}},
  \code{\link{weight.pls}}
}

